% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{extract_seq}
\alias{extract_seq}
\title{Extract the genomic sequence in LRIscan result}
\usage{
extract_seq(reference, lri_output)
}
\arguments{
\item{reference}{the reference genome in fasta format.}

\item{lri_output}{dataframe in csv format, contains at least five vairables:
Name, Ai, Ai, Bi, Bj}
}
\value{
a data frame containing extracted sequences.
}
\description{
This function takes the reference genome and LRIscan output together, and
extract the corresponding genomic sequence for further anaylsis,
}
\examples{
reference <- seqinr::read.fasta(system.file('extdata','reference_genome.fasta', package = 'LRItool'))
lri_output <- read.csv(system.file('extdata', 'lri_output.csv' , package = 'LRItool' ))
}
\author{
Haogao Gu
}
